api/requirements.txt

fastapi 
uvicorn 
numpy
requests
openai
lyzr
boto3
python-dotenv
pymongo
requests-toolbelt
python-multipart
pdfminer
pdfminer.six
databutton
pypdf
faiss-cpu
supabase
supabase_py
python-json-logger